requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.5
idna==2.10
iniconfig==1.1.1
packaging==20.8
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
